+++ lib/mini_kraken/core/goal_arg.rb
@@ -1,12 +1,10 @@
 # frozen_string_literal: true
 
-require_relative 'goal_arg'
-
 module MiniKraken
   module Core
-    # The generalization of any data value that can be
+    # The generalization of any iem that can be
     # passed as arugement to a goal.
-    class Term < GoalArg
+    class GoalArg
     end # class
   end # module
 end # module