+++ spec/support/factory_methods.rb
@@ -34,6 +34,14 @@ module MiniKraken
       Core::Goal.new(Core::Equals.instance, [arg1, arg2])
     end
 
+    # Factory method for constructing a goal using the Equals relation.
+    # @param g1 [Core::Goal]
+    # @param g2 [Core::Goal]
+    # @return [Core::Goal]
+    def conj2_goal(g1, g2)
+      Core::Goal.new(Core::Conj2.instance, [g1, g2])
+    end
+
     # Factory method for constructing a KSymbol instance
     # @param aSymbol [Symbol]
     # @return [Core::KSymbol]